requirements.txt
mock

matplotlib
numpy
pandas
scipy

Cartopy
folium
geopandas
geopy
holoviews==1.9.0
Pillow
xarray

ripe.atlas.cousteau

git+https://github.com/ioam/geoviews@1.3.2
